# cyberblue-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CyberBlue SOC API",
    "version": "1.0.0",
    "description": "Security Operations Center API for tool management and monitoring"
  },
  "servers": [
    {
      "url": "https://api.soc.local",
      "description": "Production server"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/actions/{tool_id}/{operation}": {
      "put": {
        "summary": "Perform tool action",
        "description": "Start, stop, or restart a SOC tool (Admin only)",
        "tags": ["actions"],
        "parameters": [
          {
            "name": "tool_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Tool ID"
          },
          {
            "name": "operation",
            "in": "path",
            "required": true,
            "schema": {
              "enum": ["start", "stop", "restart"],
              "type": "string"
            },
            "description": "Operation to perform"
          }
        ],
        "responses": {
          "200": {
            "description": "Action performed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "429": {
            "description": "Too Many Requests - Rate limit exceeded"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token from Keycloak authentication"
      }
    },
...
  }
}